Current attempt at NDN-NNN face
This is not going as planned. For now, I think that the only way to get the interface working is having a multimap with the NDN Name and Address maintained. This can then be returned using the multimap information.

Setting that up will be difficult. I have way more to do than I planned.