Implementation of the DU package
These are all the files that are need to be modified when you add a new packet. This is way too complicated. There has to be an easier way of adding a packet without having to modify so many things.

I will have to investigate this once I am done with the thesis.