Added MDO to all files
I have completed the addition of the MDO packet to all parts of the project.

As always, I feel that this could have been handled in another way. For now it compiles and should be enough.

I will need to update the applications and finally start working on the forwarding/routing strategy.