Consumer/Producer implementation - Names Container now uses Ptr<>
This update is huge. I realized that the Names Container was going to end up using too much RAM because of having to copy the NNNAddress constantly. I replaced and tested the new implementation using Ptr<>.

The Consumers and Producers have been adapted to use 3N names as well as NDN names. Attributes have been also used to put a lifetime for 3N based PDUs. Consumers now can decide whether to use SOs or not and Producers have a different flowchart because they can decide to use DUs or not.

The code compiles, however we are still far from finishing the prototype.