ForwardingStrategy reorganization
The file was much messier than I wanted it to be so I rearranged it and made it compile with all the information I thought was required. It is a huge update as it is basically reorganizing everything and putting it almost in its original state when it was added to the repository.

Since it compiles with the new linking data, I believe it should help when having to write the 3N code.