Implementation of PDUBuffer
This is the complete functioning PDUBuffer class. I wanted to be able to add std::queue directly to the trie.h, but I wasn't able to easily do that, forcing me to create the PDUQueue class, which ended up being a simple wrapper capable of being used with ns-3's Ptr classes. The test file show the basic functions of the class.