Heavy commit before moving
Included the PDUBuffer into ForwardingStrategy. Not sure if this is the best way to do it. For now it compiles.

Implemented the flushBuffer function in the ForwardingStrategy and included it in the OnINF information.

Modified NNNAddress to give a copy of it's internal vector via getName.

Small modifications in header to ensure compilation. I am sure that at some point some part of this will break. For now, it is ok.