Complete deletion of MDO PDU
After talking with the team, it became obvious that the MDO PDU was not required and that it was causing issues when compiling.

What is amazing is the number of files that had to be modified to remove the MDO from the code. I will probably have to refactor the code so that adding a new PDU will be easier in the future.