Fixed includes to be able to use NNNAddrAggregator class
I had to reimplement the NNNAddrEntry class within NNNAddrAggregator. I am not sure if this was entirely necessary.

I did find out that the NNNAddress class has to be included before the trie_with_policy, or you get a load of errors.

Having talked to Miyamoto today and seeing how I am adding the 3N name information to the PIT, I do not know if the NS-3 Apps need to be able to talk 3N or just ICN.

In the case of using ICN, I can use the existing code base and implement the ForwardingStrategy to use OnInterest and OnData correctly.

This will not be very fun.