PDU processing. OnX ProcessICNPDU, UpdatePITEntry, ProcessInterest and ProcessData
This is a compiling copy of the PDU selection and processing ForwardingStrategy. As far as I know, the only thing left to modify is the SatisfyPendingInterest, PropagateInterest and DoPropagateInterest.

All these forwarding strategies require modification to use the IncomingFace data. The good thing is that the PDUs have been put in a single class NNNPDU which makes the code much cleaner. I am really happy that I finally understand the inheritance in these classes.