Rough implementation of SatisfyPendingInterest using 3N
This is the first attempt to implement the function using 3N and ICN names. I am sure that it can be better summarized, but this is the best I can do at this moment.

My biggest problem was the aggregation of 3N names for sectors that were not part of the node's particular Node name.

I also implemented a simple wrapper function to test if a particular name belongs, or has belonged to a node.