NNNStackHelper initialization complete
This update is massive as it follows the the NNNStackHelper adding everything to a node.

The update includes some minor formatting errors.

The creation of the NNPT was solved by using SetTypeId in the constructor of the NNNStackHelper. After so much trouble with the code, I also cleaned up the name space for the NNPT, creating a nnpt name space for the entries. This change created a cascade of issues which were solved. The NNPT class also got a complete GetTypeId function so it could be found by the Ns-3 constructors.

The test files for the NNPT class were also changed accordingly.

The NS-3 object aggregation only works if there is one class aggregated. This means that you can't aggregate two NNPT classes. This is of special interest because the ForwardingStrategy uses two NNST classes and two NamesContainer classes. This means that only one can be aggregated. Special care must be taken in the higher layers so as to not end up looking for something that cannot be aggregated.

Finally in the ForwardingStrategy class, all non-aggregated classes that are needed, like the other NNST, the two NamesContainer and the PDUBuffer, must be initialized manually. This has been done in the constructor for the ForwardingStrategy.

As always, the code compiles. What is new in this update is that even when installing the NNNStack, the code exits successfully.