Started testing the Reenroll.
Things are not going so smoothly. The times are not in sync, making the NamesContainer fire before the NNST in the 3N name producing node fire. This created a loop that makes the producing node have too many names produced. I have to do some calculation in the code to take this into account.

I have also added code to make the logging output more understandable.

NamesContainer had callbacks that were null. Created ifs that deal with this possibility.