MultiIndexContainer sorts by less, changed to PtrNNNComp
I was having one hell of an error in the reenrolling which caused the Ptr used to compare to always be empty. After a lot of searching, I realized that the less function used in multi_index_container was being used on Ptr. Since I was always creating the Ptrs, I never got a match. Adding the comparison function and modifying the layout of the NNPT class made me get the results I needed.

The test file has also been modified to reflect the heavy Ptr usage and different Ptr holding the same value.