App has been put to connect to the the FW having or not having a name
Initially, I thought that the App having a buffer was necessary. Although useful in practice, it does not seem necessary at this stage. The complications begin as to when to use the buffer and when not to use the buffer. When you start to think in that way, it seems that the Forwarding Strategy should be taking care of that and not the Apps.

Small changes to how the name in use is propagated through the applications has also been added, making the code a little cleaner.