Testing the mobility usage for Applications
Restructured SatisfyPendingInterest. This function continues to be the most important one and is giving me headaches.

I have also tested the node using mobility.

There seems to be an issues in the App trace when using NULL PDUs that has me a little worried. Will have to look at that later.