2 Node App test with CSMA, P2P and Wifi
This has lead to the discovery that P2P links are failing. I don't know why.

I have also noticed that the PoA names in EN and REN PDUs have to be organized depending on the forwarding interface. This is a headache but will require modification to GetAllPoANames or the REN and EN creation code.

The App DEN will have to wait until these two things are solved.