FlexPointToPointNetDevice creation
I realized that the problem in the P2P links was due to the hardcoding of the protocols in the PointToPointNetDevice class. I attempted to inherit it, but I ended up having to reimplement all the classes because of the hardcoding of the Channel and RemoteChannel classes.

The result is a lot of new code that may be not needed in the long run, but that, at least, means that I don't have to modify the already installed PointToPointNetDevice classes should they exist on the target machine.

This concluded the 2 node App tests.