Use of Address class for serialization
I found that the enrollment procedures were faulty, do to the fact that I was not copying the addresses correctly. This created discrepancies when checking to see if the PoA names given where actually valid. This was mostly due to not reading about the Address types and length used by ns-3. This discovery lead to much cleaner code in the serialization section as well as the elimination of the PoA Type within the Enrollment PDUs.

In the App DEN test, the code for the forced SSID switch has been written and tested, though not completely due to the errors found in the Address class serialization.

As I have gotten better at writing NS-3 code, I now know how to toggle better logging capabilities without having to write the name of the function in every NS_LOG_INFO used. This has lead to a couple of rewrite for the logging.