SatisfyPendingInterest changes
Fixed a trace source that made it impossible for the SatisfiedInterest data in the tracer classes to be called.

SatisfyPendingInterest has been dramatically changed after testing using the nnn-2-app-den-test. It showed that the function was eating PDUs without making any judgement. The new code seems to work for the simple scenario.

In the OnSO, OnDU functions, there was a guard to protect against not enrolled machines. Although cute in theory, the PDUs to not necessarily use an enrolled name in the PDU, especially when the guard was activated in nodes that are higher up in the hiearchy.