Passed 3N name production to uint64_t style
The random, although interesting, made it exceedingly hard to follow the debugging. It doesn't help that the enrollment is working too well.

The error seems to continue in the NNST and finding the closest sector. I will probably have to create a NNST backup for when a Face cannot be selected.