Errors in REN
It seems that the Reenroll had much the same issue as Enroll in that the PDUs sometimes collided in transit, requiring a forced scheduling.

I also seem to have a error in the NNST where the longest_prefix_match is not working as required. I am not sure as to the nature of the problem but it will need to be solved for the REN process.