NNST ClosestSector
The prior version of ClosestSector choosing was based on only the longest prefix search in the trie. This meant that if you didn't have anything that was a subsector of what you were looking for then the NNST returned nothing.

This was interesting in the basic tests where you have an reenrolling node with a.0.0.0 and only a.1 and a.1.0.0 as your possible choices. As you can see, neither option is a subsector of a.0.0.0.

I seriously overlooked this problem, and is probably happening in even the DEN based tests I had done before.

Now, if you do find a subsector, then by all means that should be your next destination. If you don't, then, like in the example, you have to pick between two choices that don't seem good at first. You discriminate these choices by the distance function for NNNAddress, which should have appeared before in the NNST code.

This commit adds this code, as well as some of the test file changes made while searching for the error.