Producer side mobility has been integrated
The Consumer has been modified so that it can pick when to use the different PDU effectively. The isMobile flag is working as desired.

The forwarding strategy has been modified to take pass throughs into account.

What is of interest is that in the SatisfyPending, when responding from CS that are not the Applications, one cannot use DU PDUs. I am thinking that this could be sorted out in more complex scenarios, but for now any response that is not from an application will use DO PDUs.

This commit also includes a lot of Interest Satisfaction/Timeout processing and ends up giving coherent results.