ForwardingStrategy reimplementation
This is my attempt to fix the Producer mobility. What I have found has made the SatisfyPendingInterest function into a monster, but it has made a lot of the results far better than expected. The 5.6 m/s simulation is showing errors which I will have to look into.

I think that the issue is for ICN Data packets that do not have PIT Entries. For these, there is no propagation method, so I will have to reimplement it.

I am not sure if a buffered PDU should also not trigger the deletion of a PIT Entry. This might make it complicated to code, but I feel it is the logical solution.

This branch represents by current best attempt. However I will not merge back to the master branch until I am sure that there are no more hidden bugs.