Found bug in App code that made the App not use 3N name
If you initialize the App after enrollment, the App didn't check for name. This probably lead to more PDU loss than necessary.

The ForwardingStrategy ACK times have been lowered to 1s.

The nnn-icc-mobility scenario had a mistaken line giving the Consumer mobile conditions in both Consumer mobility and Producer mobility scenarios. This was not intended so it was fixed.